requirements.txt
asgiref==3.8.1
boto3==1.37.29
botocore==1.37.29
Django==5.1.7
django-storages==1.14.6
jmespath==1.0.1
pillow==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
